Fixing dismissal & saving when adding a new category/phrase
